refactored firestoreService, fixed totalShare bug
